requirements.txt
Django==1.8.3
django-debug-toolbar==1.3.2
django-registration-redux==1.2
pyapi-gitlab==7.8.4
requests==2.7.0
sqlparse==0.1.16
wheel==0.24.0
